"affichage des produits avec le hyperlien vers la page de details.php."

+++ produits.php
@@ -1,30 +1,37 @@
 <?php
-require_once('DB/connection.php');
-require_once ('common/defines.php');
-require_once ('views/page_top.php');
+if  (array_key_exists('sexe',$_GET) && ($_GET['sexe']==0 ||($_GET['sexe']==1))){
+    require_once('DB/connection.php');
+    require_once('common/defines.php');
+    require_once('views/page_top.php');
 
+    $sexe = $_GET['sexe'];
+    $cat_id = null;
 
-$cat_id = (isset($_GET['cat_id'])? $_GET['cat_id']: null);
-$sexe= (isset($_GET['sexe'])? $_GET['sexe']: null);
-
-$chaussures = get_chaussures_avec_cats($cat_id, $sexe);
-
-?>
-<main id="main_produits">
-    <?php
-    foreach ($chaussures as $id => $c){
-        //var_dump($chaussures);
-        $img = '<img src="'.$c["full_image_path"].'" alt= "image de chaussure"/>';
-       // var_dump($img);
-
-        echo    '<div class="chaussures">'.
-                    $c["marque"] .$c["nom"] .
-                    $img. $c["description"].
-                    $c["prix"].
-                    $c["pointure"].
-                    "<a href='?op=ajouter&itemid=".$id.((!is_null($sexe))?"&sexe=".$sexe : "")."'>Ajouter</a>".
-                '</div>';
+    if  (array_key_exists('cat_id',$_GET) && categorie_exists($_GET['cat_id'])) {
+        $cat_id = $_GET['cat_id'];
     }
+
+    $chaussures = get_chaussures_avec_cats($cat_id, $sexe);
+
     ?>
-</main>
-<?php require_once ('views/page_bottom.php') ?>
+
+    <main id="main_produits">
+        <?php
+        foreach ($chaussures as $id => $c){
+            $img = '<a href="details.php?produit_id=' . $c['id']. '"><img src="'.$c["full_image_path"].'" alt= "image de chaussure"/></a>';
+
+            echo    '<div class="chaussures">'.
+                ucfirst($c["marque"] ). ' ' .ucfirst($c["nom"]) .
+                $img.
+                $c["prix"]. ' '.
+                "<a href='?op=ajouter&itemid=".$id.((!is_null($sexe))?"&sexe=".$sexe : "")."'>Ajouter</a>".
+                '</div>';
+        }
+        ?>
+    </main>
+    <?php require_once('views/page_bottom.php') ?>
+    <?php
+}else {
+   header('Location: index.php');
+}
+?>
